@@ -4,19 +4,19 @@
   "private": true,
   "proxy": "http://localhost:5000",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "concurrently": "^6.3.0",
-    "framer-motion": "^4.1.17",
+    "@testing-library/jest-dom": "^5.16.3",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^14.0.1",
+    "concurrently": "^7.0.0",
+    "framer-motion": "^6.2.8",
     "json-server": "^0.17.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-icons": "^4.3.1",
-    "react-router-dom": "^6.0.1",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.2.2",
+    "react-scripts": "5.0.0",
     "uuid": "^8.3.2",
-    "web-vitals": "^1.1.2"
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
